chore: Move _validate_datestring to defence.py
