+++ src/transport_performance/gtfs/gtfs_utils.py
@@ -10,50 +10,14 @@ import plotly.graph_objects as go
 from typing import Union
 import pathlib
 from geopandas import GeoDataFrame
-from datetime import datetime
 
 from transport_performance.utils.defence import (
     _is_expected_filetype,
     _check_iterable,
     _type_defence,
+    _validate_datestring,
 )
 from transport_performance.utils.constants import PKG_PATH
-
-
-def _validate_datestring(date_text: str, form: str = "%Y%m%d") -> None:
-    """Ensure a date string conforms to the expected form.
-
-    Parameters
-    ----------
-    date_text : str
-        The datestring to be checked.
-    form : str, optional
-        The expected date format. Must be a valid C-standard date code format
-        compatible with datetime. See https://docs.python.org/3/library/datetim
-        e.html#strftime-and-strptime-format-codes for examples. Defaults to
-        "%Y%m%d".
-
-    Returns
-    -------
-    None
-
-    Raises
-    ------
-    ValueError
-        `date_text` is not in the expected `form`.
-    TypeError
-        `date_text` or `form` are not string.
-
-    """
-    _type_defence(date_text, "date_text", str)
-    _type_defence(form, "form", str)
-    try:
-        datetime.strptime(date_text, form)
-    except ValueError:
-        raise ValueError(
-            f"Incorrect date format, {date_text} should be {form}"
-        )
-    return None
 
 
 def bbox_filter_gtfs(
