+++ src/transport_performance/utils/defence.py
@@ -7,6 +7,7 @@ import numpy as np
 import os
 import pandas as pd
 import warnings
+from datetime import datetime
 
 
 def _handle_path_like(
@@ -488,3 +489,39 @@ def _enforce_file_extension(
         warnings.warn(msg, UserWarning)
         path = os.path.normpath(root + default_ext)
     return pathlib.Path(path)
+
+
+def _validate_datestring(date_text: str, form: str = "%Y%m%d") -> None:
+    """Ensure a date string conforms to the expected form.
+
+    Parameters
+    ----------
+    date_text : str
+        The datestring to be checked.
+    form : str, optional
+        The expected date format. Must be a valid C-standard date code format
+        compatible with datetime. See https://docs.python.org/3/library/datetim
+        e.html#strftime-and-strptime-format-codes for examples. Defaults to
+        "%Y%m%d".
+
+    Returns
+    -------
+    None
+
+    Raises
+    ------
+    ValueError
+        `date_text` is not in the expected `form`.
+    TypeError
+        `date_text` or `form` are not string.
+
+    """
+    _type_defence(date_text, "date_text", str)
+    _type_defence(form, "form", str)
+    try:
+        datetime.strptime(date_text, form)
+    except ValueError:
+        raise ValueError(
+            f"Incorrect date format, {date_text} should be {form}"
+        )
+    return None
